[PATCH/ADD/REFACTOR] Major changes including adding a simplex container, using C++17 standard now, adding first version of expression templates and changing some design patterns.

+++ cpp/mesh/include/types.h
@@ -9,7 +9,7 @@
 
 typedef unsigned long long int ullong;
 
-typedef int ID;
+typedef long ID;
 
 using EigenDStride = Eigen::Stride<Eigen::Dynamic, Eigen::Dynamic>;
 template<typename MatrixType> using EigenDRef = Eigen::Ref<MatrixType, 0, EigenDStride>;
